requirements.txt
# ===============================
# Leak Detection Project - Requirements
# ===============================

# --- Core Frameworks ---
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# --- Diffusion Models (Stable Diffusion) ---
diffusers==0.28.0
transformers==4.41.2
accelerate==0.30.1
safetensors==0.4.3

# --- Image Processing and Data Handling ---
opencv-python==4.10.0.84
pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2

# --- Visualization and Explainability ---
grad-cam==1.5.3
tensorboard==2.16.2

# --- Environment & Utilities ---
python-dotenv==1.0.1
requests==2.32.3

# --- Optional (for HPC / ONNX export) ---
onnx==1.16.0
onnxruntime==1.17.1
